feat(mock-page-ui): add checkout page
